// store/seller/seller.types.ts
enum SellerTypes {
  SELLER_VERIFICATION = "SELLER_VERIFICATION",
  COMPLETE_SELLER_VERIFICATION = "COMPLETE_SELLER_VERIFICATION",
  START_SELLING = "START_SELLING",
  GET_SELLER = "GET_SELLER",
  SELLER_LOADING = "SELLER_LOADING",
  SELLER_ERROR = "SELLER_ERROR",
  CLEAR_SELLER = "CLEAR_SELLER",
}

export default SellerTypes;